commit save

+++ Assets/XCharts-master/Runtime/XLog/XLog.cs
@@ -8,8 +8,8 @@ using UnityEngine;
 namespace XCharts.Runtime
 {
     /// <summary>
-    /// Log system. Used to output logs with date and log type, support output to file, support custom output log type.
-    /// ||日志系统。用于输出带日期和日志类型的日志，支持输出到文件，支持自定义输出的日志类型。
+    ///     Log system. Used to output logs with date and log type, support output to file, support custom output log type.
+    ///     ||日志系统。用于输出带日期和日志类型的日志，支持输出到文件，支持自定义输出的日志类型。
     /// </summary>
     public class XLog : MonoBehaviour
     {
@@ -24,70 +24,47 @@ namespace XCharts.Runtime
 
         private const int MAX_ERROR_LOG = 20;
 
-        public static bool isReportBug = false;
+        public static bool isReportBug;
         public static bool isOutputLog = false;
-        public static bool isUploadLog = false;
+        public static bool isUploadLog;
         public static bool isCloseOutLog = false;
 
-        public static int errorCount = 0;
-        public static int exceptCount = 0;
+        public static int errorCount;
+        public static int exceptCount;
         public static int uploadTick = 20;
         public static int reportTick = 10;
 
-        private static bool initFileSuccess = false;
-        private static bool[] levelList = new bool[] { true, true, true, true, true, true, true, true };
-        private static List<string> writeList = new List<string>();
-        private static float uploadTime = 0;
-        private static float reportTime = 0;
+        private static bool initFileSuccess;
+        private static readonly bool[] levelList = { true, true, true, true, true, true, true, true };
+        private static readonly List<string> writeList = new();
+        private static float uploadTime;
+        private static float reportTime;
+        public static List<string> errorList = new();
+        private static readonly object m_Lock = new();
+
+        public int logCount;
 
         private string outpath;
-        private StreamWriter writer;
         private string[] temp;
+        private StreamWriter writer;
 
-        public int logCount = 0;
-        public static List<string> errorList = new List<string>();
-        private static object m_Lock = new object();
+        public static XLog Instance { get; private set; }
 
-        private static XLog m_Instance;
-        public static XLog Instance
+        private void Awake()
         {
-            get
-            {
-                // if (m_Instance == null)
-                // {
-                //     GameObject go = new GameObject("XLog");
-                //     m_Instance = go.AddComponent<XLog>();
-                //     DontDestroyOnLoad(go);
-                // }
-                return m_Instance;
-            }
-        }
-
-        void Awake()
-        {
-            if (m_Instance != null)
+            if (Instance != null)
             {
                 Destroy(gameObject);
                 return;
             }
-            m_Instance = this;
+
+            Instance = this;
             InitLogFile();
             // Application.logMessageReceived += HandleLog;
             Application.logMessageReceivedThreaded += HandleLog;
         }
 
-        void OnDestroy()
-        {
-            if (writer != null)
-            {
-                writer.Close();
-                writer.Dispose();
-            }
-            // Application.logMessageReceived -= HandleLog;
-            Application.logMessageReceivedThreaded -= HandleLog;
-        }
-
-        void Update()
+        private void Update()
         {
             uploadTime += Time.deltaTime;
             reportTime += Time.deltaTime;
@@ -101,10 +78,11 @@ namespace XCharts.Runtime
                         writeList.Clear();
                         return;
                     }
+
                     try
                     {
                         temp = writeList.ToArray();
-                        int count = 0;
+                        var count = 0;
                         foreach (var str in temp)
                         {
                             count++;
@@ -112,6 +90,7 @@ namespace XCharts.Runtime
                             writeList.Remove(str);
                             if (count > 10) break;
                         }
+
                         writer.Flush();
                     }
                     catch (Exception e)
@@ -125,30 +104,40 @@ namespace XCharts.Runtime
             }
         }
 
+        private void OnDestroy()
+        {
+            if (writer != null)
+            {
+                writer.Close();
+                writer.Dispose();
+            }
+
+            // Application.logMessageReceived -= HandleLog;
+            Application.logMessageReceivedThreaded -= HandleLog;
+        }
+
         private void InitLogFile()
         {
             ClearAllLog();
-            XLog.EnableLog(ALL);
+            EnableLog(ALL);
             if (Application.platform == RuntimePlatform.Android || Application.platform == RuntimePlatform.IPhonePlayer)
             {
-                XLog.ClearAllLog();
-                XLog.EnableLog(VITAL);
-                XLog.EnableLog(ERROR);
-                XLog.isReportBug = true;
-                XLog.isUploadLog = true;
+                ClearAllLog();
+                EnableLog(VITAL);
+                EnableLog(ERROR);
+                isReportBug = true;
+                isUploadLog = true;
             }
             else
             {
-                XLog.isUploadLog = false;
-                XLog.isReportBug = false;
+                isUploadLog = false;
+                isReportBug = false;
             }
+
             outpath = GetLogOutputPath();
             try
             {
-                if (File.Exists(outpath))
-                {
-                    File.Delete(outpath);
-                }
+                if (File.Exists(outpath)) File.Delete(outpath);
                 writer = new StreamWriter(outpath, false, Encoding.UTF8);
                 writer.WriteLine(GetNowTime() + "init file success!!");
                 UnityEngine.Debug.Log(GetNowTime() + "init file success:" + outpath);
@@ -166,7 +155,7 @@ namespace XCharts.Runtime
         private static string GetLogOutputPath()
         {
 #if UNITY_EDITOR
-            string path = Application.dataPath + "/../outlog.txt";
+            var path = Application.dataPath + "/../outlog.txt";
 #else
             string path = Application.persistentDataPath + "/outlog.txt";
             if (Application.platform == RuntimePlatform.Android || Application.platform == RuntimePlatform.IPhonePlayer)
@@ -186,11 +175,11 @@ namespace XCharts.Runtime
             lock (m_Lock)
             {
                 if (!initFileSuccess) return;
-                int index = logString.IndexOf("stack traceback");
+                var index = logString.IndexOf("stack traceback");
                 if (index > 0)
                 {
-                    string log = logString.Substring(0, index);
-                    string trace = logString.Substring(index, logString.Length - index);
+                    var log = logString.Substring(0, index);
+                    var trace = logString.Substring(index, logString.Length - index);
                     logString = log;
                     stackTrace = trace;
                 }
@@ -206,16 +195,10 @@ namespace XCharts.Runtime
                     writeList.Add(logString);
                     //writeList.Add(stackTrace + "\n");
 
-                    if (errorList.Count >= MAX_ERROR_LOG)
-                    {
-                        errorList.RemoveAt(1);
-                    }
+                    if (errorList.Count >= MAX_ERROR_LOG) errorList.RemoveAt(1);
 
-                    if (errorList.Count < MAX_ERROR_LOG)
-                    {
-                        errorList.Add(logString);
-                       // errorList.Add(stackTrace + "\n");
-                    }
+                    if (errorList.Count < MAX_ERROR_LOG) errorList.Add(logString);
+                    // errorList.Add(stackTrace + "\n");
                 }
                 else if (type == LogType.Exception)
                 {
@@ -224,10 +207,7 @@ namespace XCharts.Runtime
                     writeList.Add(logString);
                     writeList.Add(stackTrace + "\n");
 
-                    if (errorList.Count >= MAX_ERROR_LOG)
-                    {
-                        errorList.RemoveAt(1);
-                    }
+                    if (errorList.Count >= MAX_ERROR_LOG) errorList.RemoveAt(1);
 
                     if (errorList.Count < MAX_ERROR_LOG)
                     {
@@ -240,13 +220,10 @@ namespace XCharts.Runtime
 
         public static void FlushLog()
         {
-            var instance = XLog.Instance;
+            var instance = Instance;
             if (instance != null && instance.writer != null)
             {
-                for (int i = 0; i < writeList.Count; i++)
-                {
-                    instance.writer.WriteLine(writeList[i]);
-                }
+                for (var i = 0; i < writeList.Count; i++) instance.writer.WriteLine(writeList[i]);
                 instance.writer.Flush();
                 writeList.Clear();
             }
@@ -260,10 +237,7 @@ namespace XCharts.Runtime
 
         public static void ClearAllLog()
         {
-            for (int i = 0; i < levelList.Length; i++)
-            {
-                levelList[i] = false;
-            }
+            for (var i = 0; i < levelList.Length; i++) levelList[i] = false;
         }
 
         public static bool CanLog(int level)
@@ -325,11 +299,10 @@ namespace XCharts.Runtime
 
         public static string GetNowTime(string formatter = null)
         {
-            DateTime now = DateTime.Now;
+            var now = DateTime.Now;
             if (formatter == null)
                 return now.ToString("[HH:mm:ss fff]", DateTimeFormatInfo.InvariantInfo);
-            else
-                return now.ToString(formatter, DateTimeFormatInfo.InvariantInfo);
+            return now.ToString(formatter, DateTimeFormatInfo.InvariantInfo);
         }
 
         public static ulong GetTimestamp()
